Finished all requests

+++ src/main.py
@@ -1,10 +1,9 @@
-import datetime
 import uuid
-from typing import List
-
 import uvicorn
-from fastapi import FastAPI, UploadFile, File, HTTPException
 
+from datetime import datetime
+from typing import List
+from fastapi import FastAPI, UploadFile, File, HTTPException
 from image import Image, ImageOut
 from minio_connection import MinIO
 from src.db_connection import AppDb
@@ -12,7 +11,13 @@ from src.db_connection import AppDb
 app = FastAPI()
 db = AppDb()
 min_io = MinIO()
+
 TIME_FORMAT = '%Y-%m-%d %H:%M:%S'
+
+
+def get_images(request_code):
+    return [Image(img[0], img[1], img[2], img[3].strftime(TIME_FORMAT)).__dict__
+            for img in db.get_images(request_code)]
 
 
 @app.post("/frames/", response_model=List[ImageOut])
@@ -20,7 +25,7 @@ async def upload_images(files: List[UploadFile] = File(...)):
     try:
         request_code = uuid.uuid4()
         for img in files:
-            now_time = datetime.datetime.now()
+            now_time = datetime.now()
             bucket_name = f'{now_time.year}{now_time.month}{now_time.day}'
             contents = await img.read()
             file_name = f"{uuid.uuid4()}.jpeg"
@@ -28,8 +33,7 @@ async def upload_images(files: List[UploadFile] = File(...)):
             min_io.upload_image(bucket_name, file_name, contents)
             db.insert_image(Image(None, request_code, file_name, now_time.strftime(TIME_FORMAT)))
 
-        reply = [Image(img[0], img[1], img[2], img[3].strftime(TIME_FORMAT)).__dict__
-                 for img in db.get_images(request_code)]
+        reply = get_images(request_code)
 
     except HTTPException:
         return {'message': 'There was an error uploading the file(s)'}
@@ -40,8 +44,7 @@ async def upload_images(files: List[UploadFile] = File(...)):
 @app.get("/frames/{request_code}", response_model=List[ImageOut])
 async def read_images(request_code: str):
     try:
-        return [Image(img[0], img[1], img[2].strftime(TIME_FORMAT)).__dict__
-                for img in db.get_images(request_code)]
+        return get_images(request_code)
 
     except HTTPException:
         return {'message': 'There was an error reading the file(s)'}
@@ -50,10 +53,16 @@ async def read_images(request_code: str):
 @app.delete("/frames/{request_code}")
 async def delete_images(request_code: str):
     try:
-        images_to_delete = [Image(img[0], img[1], img[2].strftime(TIME_FORMAT)).__dict__
-                            for img in db.get_images(request_code)]
+        images_to_delete = [Image(img[0], img[1], img[2], img[3]) for img in db.get_images(request_code)]
+        if not images_to_delete:
+            return {'message': 'No files with such request code'}
+        else:
+            for image in images_to_delete:
+                min_io.delete_image(f'{image.reg_date.year}{image.reg_date.month}{image.reg_date.day}', image.file_name)
 
-        db.delete_images(request_code)
+            db.delete_images(request_code)
+
+        return {'message': f'{len(images_to_delete)} file(s) were successfully deleted'}
 
     except HTTPException:
         return {'message': 'There was an error deleting the file(s)'}
